requirements.txt
torch==2.0.1
torchvision
diffusers==0.19.3
transformers==4.36.0
xformers==0.0.20
xatlas
accelerate
cupy-cuda11x==12.2.0
configargparse
ipython
fvcore
trimesh
embreex
pandas
sigal

# LLM
openai
python-dotenv